+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Basic parsing and folding of Docker build stage #400
 - Shorter names for utility steps #401
 - Failing utility doesn't create blank cache #402
+- Allow project list API meta on filtered queries #404
 
 ### v0.0.9
 - Exception reported to Rollbar on git push hook `ValidationError` #322